+++ packages/admin/lib/router.dart
@@ -81,6 +81,10 @@ Route<dynamic> generateRoute(RouteSettings settings) {
       return MaterialPageRoute(
           builder: (context) => DisplayMenuOption(
               menuList: acctMenuOptions, menuIndex: 3, tabIndex: 0));
+    case '/acctReports':
+      return MaterialPageRoute(
+          builder: (context) => DisplayMenuOption(
+              menuList: acctMenuOptions, menuIndex: 4, tabIndex: 0));
     default:
       return coreRoute(settings);
   }